Working on adding retina support
This involves heavily modifying the KISS_SDL library, which may break
some stuff. Hopefully we can make a pull request with this support to
the library itself, when it’s working.

Work on this feature has already begun; this branch should have been
created a while ago. For full commit history, please see the master
branch.